lidt ændringer til carporten

+++ Code-carport/src/main/java/DBAccess/CarportMapper.java
@@ -13,9 +13,9 @@ public class CarportMapper {
             Connection con = Connector.connection();
             String SQL = "INSERT INTO Carport (length, width, height) VALUES (?, ?, ?)";
             PreparedStatement ps = con.prepareStatement(SQL, Statement.RETURN_GENERATED_KEYS);
-            ps.setDouble(1, carport.getLength());
-            ps.setDouble(2, carport.getWidth());
-            ps.setDouble(3, carport.getHeight());
+            ps.setInt(1, carport.getLength());
+            ps.setInt(2, carport.getWidth());
+            ps.setInt(3, carport.getHeight());
             ps.executeUpdate();
 
         } catch (ClassNotFoundException | SQLException e) {
